[#40] Provide cleaner contracts
