+++ cache_test.go
@@ -22,23 +22,17 @@ import (
 	"testing"
 	"time"
 
-	"github.com/maypok86/otter/internal/node"
 	"github.com/maypok86/otter/internal/xruntime"
 )
 
 func TestCache_Set(t *testing.T) {
-	b, err := NewBuilder[int, int](100)
-	if err != nil {
-		t.Fatalf("can not create builder: %v", err)
-	}
-
-	c, err := b.StatsEnabled(true).Build()
+	c, err := MustBuilder[int, int](100).WithTTL(time.Minute).CollectStats().Build()
 	if err != nil {
 		t.Fatalf("can not create cache: %v", err)
 	}
 
 	for i := 0; i < 100; i++ {
-		c.SetWithTTL(i, i, time.Minute)
+		c.Set(i, i)
 	}
 
 	parallelism := xruntime.Parallelism()
@@ -68,7 +62,7 @@ func TestCache_Set(t *testing.T) {
 	if err != nil {
 		t.Fatalf("not found key: %v", err)
 	}
-	ratio := c.Ratio()
+	ratio := c.Stats().Ratio()
 	if ratio != 1.0 {
 		t.Fatalf("cache hit ratio should be 1.0, but got %v", ratio)
 	}
@@ -76,13 +70,13 @@ func TestCache_Set(t *testing.T) {
 
 func TestCache_SetWithTTL(t *testing.T) {
 	size := 256
-	c, err := MustBuilder[int, int](size).Build()
+	c, err := MustBuilder[int, int](size).WithTTL(time.Second).Build()
 	if err != nil {
 		t.Fatalf("can not create builder: %v", err)
 	}
 
 	for i := 0; i < size; i++ {
-		c.SetWithTTL(i, i, time.Second)
+		c.Set(i, i)
 	}
 
 	time.Sleep(3 * time.Second)
@@ -98,65 +92,35 @@ func TestCache_SetWithTTL(t *testing.T) {
 		t.Fatalf("c.Size() = %d, want = %d", cacheSize, 0)
 	}
 
-	c, err = MustBuilder[int, int](size).Build()
+	cc, err := MustBuilder[int, int](size).WithVariableTTL().CollectStats().Build()
 	if err != nil {
 		t.Fatalf("can not create builder: %v", err)
 	}
 
 	for i := 0; i < size; i++ {
-		c.SetWithTTL(i, i, 5*time.Second)
+		cc.Set(i, i, 5*time.Second)
 	}
 
 	time.Sleep(7 * time.Second)
 
 	for i := 0; i < size; i++ {
-		if c.Has(i) {
+		if cc.Has(i) {
 			t.Fatalf("key should be expired: %d", i)
 		}
 	}
 
 	time.Sleep(10 * time.Millisecond)
 
-	if cacheSize := c.Size(); cacheSize != 0 {
+	if cacheSize := cc.Size(); cacheSize != 0 {
 		t.Fatalf("c.Size() = %d, want = %d", cacheSize, 0)
 	}
-}
-
-func TestCache_SetWithCost(t *testing.T) {
-	size := 10
-	c, err := MustBuilder[int, int](size).Cost(func(key int, value int) uint32 {
-		return uint32(key)
-	}).Build()
-	if err != nil {
-		t.Fatalf("can not create cache: %v", err)
-	}
-
-	goodCost := int(c.policy.MaxAvailableCost())
-	badCost := goodCost + 1
-
-	added := c.Set(goodCost, 1)
-	if !added {
-		t.Fatalf("Set was dropped, even though it shouldn't have been. Max available cost: %d, actual cost: %d",
-			c.policy.MaxAvailableCost(),
-			c.costFunc(goodCost, 1),
-		)
-	}
-	added = c.Set(badCost, 1)
-	if added {
-		t.Fatalf("Set wasn't dropped, though it should have been. Max available cost: %d, actual cost: %d",
-			c.policy.MaxAvailableCost(),
-			c.costFunc(badCost, 1),
-		)
+	if misses := cc.Stats().Misses(); misses != int64(size) {
+		t.Fatalf("c.Stats().Misses() = %d, want = %d", misses, size)
 	}
 }
 
 func TestCache_Ratio(t *testing.T) {
-	b, err := NewBuilder[uint64, uint64](100)
-	if err != nil {
-		t.Fatalf("can not create builder: %v", err)
-	}
-
-	c, err := b.StatsEnabled(true).Build()
+	c, err := MustBuilder[uint64, uint64](100).CollectStats().Build()
 	if err != nil {
 		t.Fatalf("can not create cache: %v", err)
 	}
@@ -174,97 +138,7 @@ func TestCache_Ratio(t *testing.T) {
 	}
 
 	t.Logf("actual size: %d, capacity: %d", c.Size(), c.Capacity())
-	t.Logf("actual: %.2f, optimal: %.2f", c.Ratio(), o.Ratio())
-}
-
-func TestCache_Range(t *testing.T) {
-	size := 10
-	c, err := MustBuilder[int, int](size).Build()
-	if err != nil {
-		t.Fatalf("can not create cache: %v", err)
-	}
-
-	time.Sleep(3 * time.Second)
-
-	c.Set(1, 1)
-	c.hashmap.Set(node.New(2, 2, 1, 1))
-	c.Set(3, 3)
-	aliveNodes := 2
-	iters := 0
-	c.Range(func(key, value int) bool {
-		if key != value {
-			t.Fatalf("got unexpected key/value for iteration %d: %d/%d", iters, key, value)
-			return false
-		}
-		iters++
-		return true
-	})
-	if iters != aliveNodes {
-		t.Fatalf("got unexpected number of iterations: %d", iters)
-	}
-}
-
-func TestCache_Close(t *testing.T) {
-	size := 10
-	c, err := MustBuilder[int, int](size).Build()
-	if err != nil {
-		t.Fatalf("can not create cache: %v", err)
-	}
-
-	for i := 0; i < size; i++ {
-		c.Set(i, i)
-	}
-
-	if cacheSize := c.Size(); cacheSize != size {
-		t.Fatalf("c.Size() = %d, want = %d", cacheSize, size)
-	}
-
-	c.Close()
-
-	time.Sleep(10 * time.Millisecond)
-
-	if cacheSize := c.Size(); cacheSize != 0 {
-		t.Fatalf("c.Size() = %d, want = %d", cacheSize, 0)
-	}
-	if !c.isClosed {
-		t.Fatalf("cache should be closed")
-	}
-
-	c.Close()
-
-	if cacheSize := c.Size(); cacheSize != 0 {
-		t.Fatalf("c.Size() = %d, want = %d", cacheSize, 0)
-	}
-	if !c.isClosed {
-		t.Fatalf("cache should be closed")
-	}
-}
-
-func TestCache_Clear(t *testing.T) {
-	size := 10
-	c, err := MustBuilder[int, int](size).Build()
-	if err != nil {
-		t.Fatalf("can not create cache: %v", err)
-	}
-
-	for i := 0; i < size; i++ {
-		c.Set(i, i)
-	}
-
-	if cacheSize := c.Size(); cacheSize != size {
-		t.Fatalf("c.Size() = %d, want = %d", cacheSize, size)
-	}
-
-	c.Clear()
-
-	time.Sleep(10 * time.Millisecond)
-
-	if cacheSize := c.Size(); cacheSize != 0 {
-		t.Fatalf("c.Size() = %d, want = %d", cacheSize, 0)
-	}
-	if c.isClosed {
-		t.Fatalf("cache shouldn't be closed")
-	}
+	t.Logf("actual: %.2f, optimal: %.2f", c.Stats().Ratio(), o.Ratio())
 }
 
 type optimal struct {
